#17 Property and Type are now part of the foundation; errors and validations moved to the top-level.
