+++ cml-compiler/cml-language/src/main/java/cml/language/errors/ErrorReport.java
@@ -1,4 +1,4 @@
-package cml.language.foundation.errors;
+package cml.language.errors;
 
 import cml.language.foundation.NamedElement;
