+++ cml-compiler/cml-language/src/main/java/cml/language/features/Target.java
@@ -2,6 +2,7 @@ package cml.language.features;
 
 import cml.language.foundation.ModelElement;
 import cml.language.foundation.NamedElement;
+import cml.language.foundation.PropertyList;
 import cml.language.foundation.Scope;
 
 import java.util.List;
